VERSA/requirements.txt
langchain==0.1.19
streamlit==1.34.0
tqdm==4.66.4
graphviz==0.20.3
langchain_openai==0.1.6
python-pptx==0.6.23
pandasql==0.7.3
snowflake-connector-python==3.10.0
tabulate==0.9.0
gnews==0.3.7
beautifulsoup4==4.12.3
fake-useragent==1.5.1